Add files via upload
This one I owe to Liz Rice. Thanks to her incredible presentation at GOTO 2018, I was able to understand what containers even are and how they work. Taking a peek under the hood of docker and answering the question: just what is a container?

This code, written in Go, creates a container that runs in three new namespaces (UTS, PIDS and NS), renames the container's hostname, uses a new specified root directory, and limits the pids to 20.

While I typed out every character of this code manually, it all belongs to Liz! I'm just posting it here to record my journey in learning cloud infrastructure.